Major refactor/rewrite.
The whole library is essentialy rewritten.

In addition to the primary, high level API traits, it also exposes
lower level "raw" forms of those API traits, allowing users to hook
into the library at a lower level. This feature is not fully formed
yet.

Relationship updating traits have been added, but especially for the
to-many relations, these are current a bit broken.

There's now a step of disjointness between resources and relations,
with the new Relation trait, allowing e.g. the Author relation, for
which the resouce is a User.

A lot is still very wrong, but this feels like a much firmer
foundation for future work.