Use Entities and Reprs to improve efficiency.
**Entity**

Entity is an enum which could either by a resource's Id, or that
resource itself. This is passed to each rel method, instead of the id,
so that the result possible queries performed during the resource
method can be shared between them.

For example, during a GET request, the `Get` method is called, then
the `HasOne` and `HasMany` impls for each of that resource's relations
are called. If the resource and its related resources are backed by
database records, the ids of that resource's relations can be fetched
in the same query used to fetch that resource's attributes, and those
ids can be stored in the Self type to be used in the relation methods.

**Repr**

The Resource type now has an associated type Repr. Whereas before, the
Resource was required to implement Serialize, the Repr type now must
implement Serialize (and, in the case of Post resources, Deserialize).

When presenting the resource in the API, it is presented by serializing
its Repr value, not the resource itself. This can make it easier to
distinguish the API representation from the internal representation of
the resource.