First draft of Post/Patch Async.
If the processing performed on a POST or PATCH takes too long, the user
can implement PostAsync or PatchAsync instead. These return an AsyncJob
type instead of the resource; the AsyncJob is a resource at which info
about the processing can be found.

Currently, AsyncJob is expected to implement a way to accept relations
to be linked once the job is completed, plus some JSON API traits.

Very little support is provided for making sure the AsyncJob's behavior
is spec compliant. More support will hopefully be added in the future.