Use trait object for includes.
Prior to this commit, the attributes in an Include struct were
represented as a json::Value. This was to eliminate the Resource type
parameter so that includes of multiple resources could be combined in a
single collection.

By using changes to serde on an experimental branch, we are able to
represent the include attributes using a boxed trait instead. This is
significantly more efficient than pre-serializing them to json::Values.