Add Represent and Presenter.
In order to support sparse field sets, introduce a new trait. This
trait is essentially a special form of Serialize; the difference is
that it is scoped to only take serializers which also implement a
trait called Presenter. Presenter has a method, field_set, which
returns the sparse field set requested.

Soon, a derive macro will be written for Represent, so that users
can derive it instead of implementing it.