Add Get Aliases.
Allow aliases of GET requests.

These aliases are intended to act as the entry point to your
hypermedia API; for example, if you have a shopping app, you might
have an alias called "cart" to fetch this user's shopping cart
resource, from which point all traversal would be driven by the
hypermedia URLs returned in the response body.

Alias methods take an AliasRequest object as an argument, which
currently only has the alias route used in the alias. In the future,
other members will be added to this struct, including items which can
be used for authentication to route different users to different
resources as their entry point response.

Aliases of other methods forthcoming as well.