Add Receiver layer.
Similar to the presenter layer, the receiver layer takes a Read object
for the body of an API endpoint request and parses it according to some
format to become an update representation. A receiver layer is
implemented for the JSON API format.

Now, neither the api layer nor the server layer know anything about
Deserialize, just as they know nothing about Serialize/Represent. Only
the JSON API presenter/receiver impose that constraint.