Complete rewrite based on tokio.
After going through 3 or 4 drafts of this project, this might be the
final version.

This is still very incomplete & has several extremely hacky elements.

The crate now has several subcrates checked in:

- mainsail is the core resource & method abstractions, depends only on
futures & not tokio or hyper.
- rigging contains the format (presenter/receiver) abstractions as
well as a lot of glue code, it depends on tokio and hyper.
- compass-rose is the new routes! macro code generator. Using the
proc-macro-hack, it passes the routes! macro to a lalrpop parser to
generate code.
- simple-debug is a basic presenter which just debug prints resources.

Still to do:

- build up all the features again.
- write the JSON API format
- write a templated html format
- making the routing code robust
- a million more things